Align authors display with docs.asp.net
We align to docs.asp.net way of treating authors by including
authors.txt file which contains the names and GitHub profile links
for all authors on the site. This replaces the previous image + bio
at the end of the pages approach. Also remove the _authors folder
since we don't need it any more.